requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
django-cleanup==9.0.0
django-modeltranslation==0.19.11
django-rosetta==0.10.1
idna==3.10
pillow==11.0.0
polib==1.2.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.2
urllib3==2.2.3
